requirements.txt
fbm==0.2.0
gym==0.12.1
pytorch-ignite==0.1.2
iisignature==0.23
jupyter==1.0.0
matplotlib==2.2.4
pandas==0.24.2
torch==1.0.1
scikit-learn==0.20.3
sdepy==1.0.1
tqdm==4.31.1

